feat: Implement popup button UI (resolve #2)
This commit adds a button in the center using a content-script.

All CSS is isolated from the site using Shadow DOM.
It is wrapped with a custom tag name from the .env file.

The imagetools plugin is customized to work without type errors.
Its contents are added to the d.ts file.

Additionally, the button's image is resized and encoded in base64 using the imagetools tool.
This is done before the image is added to the button.